+++ packages/api/package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@math3d/mathitem-configs": "workspace:^",
     "@tanstack/react-query": "^5.8.4",
-    "axios": "^1.6.5",
+    "axios": "^1.6.7",
     "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
