+++ packages/app-tests-e2e/package.json
@@ -11,12 +11,13 @@
     "@math3d/api": "workspace:^",
     "@math3d/mock-api": "workspace:^",
     "@playwright/test": "^1.41.0",
+    "axios": "^1.6.7",
     "dotenv": "^16.3.1",
+    "eml-parse-js": "^1.1.13",
     "envalid": "^8.0.0",
     "mathlive": "^0.98.6",
+    "p-retry": "^6.2.0",
     "three": "^0.147.0",
-    "tiny-invariant": "^1.3.1",
-    "eml-parse-js": "^1.1.13",
-    "p-retry": "^6.2.0"
+    "tiny-invariant": "^1.3.1"
   }
 }
