+++ packages/app-tests-e2e/src/fixtures/users.ts
@@ -1,34 +1,13 @@
 import { test as base } from "@playwright/test";
 import type { Page } from "@playwright/test";
-import { AuthApi, API_TOKEN_KEY, ScenesApi } from "@math3d/api";
+import { API_TOKEN_KEY, ScenesApi } from "@math3d/api";
 import type { Scene } from "@math3d/api";
 import invariant from "tiny-invariant";
 import env from "@/env";
-import { getConfig } from "@/utils/api";
+import { getConfig } from "@/utils/api/config";
+import { getAuthToken, users } from "@/utils/api/auth";
 
 const TEST_SCENE_PREFIX = "[TEST-E2E-SCENE]";
-
-const authApi = new AuthApi(undefined, env.TEST_API_URL);
-
-const users = {
-  /**
-   * This admin user should not be necessary in actual tests, but is useful for
-   * setup and teardown. E.g., this user is used to delete ephemeral test
-   * accounts if they exists.
-   */
-  admin: {
-    email: env.TEST_USER_ADMIN_EMAIL,
-    password: env.TEST_USER_ADMIN_PASSWORD,
-  },
-  static: {
-    email: env.TEST_USER_1_EMAIL,
-    password: env.TEST_USER_1_PASSWORD,
-  },
-  dynamic: {
-    email: env.TEST_USER_2_EMAIL,
-    password: env.TEST_USER_2_PASSWORD,
-  },
-};
 
 type Fixtures = {
   user: keyof typeof users | null;
@@ -39,18 +18,10 @@ type Fixtures = {
 
 const test = base.extend<Fixtures>({
   user: null,
-  authToken: async ({ user: userString }, use) => {
+  authToken: async ({ user }, use) => {
     let authToken: string | null = null;
-    if (userString) {
-      const user = users[userString];
-      const response = await authApi.authTokenLoginCreate({
-        TokenCreateRequest: {
-          email: user.email,
-          password: user.password,
-        },
-      });
-      // @ts-expect-error drf-spectacular issue
-      authToken = response.data.auth_token;
+    if (user) {
+      authToken = await getAuthToken(user);
     }
     await use(authToken);
   },
