+++ packages/app/package.json
@@ -48,7 +48,7 @@
     "@tanstack/react-query": "^5.8.4",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^20.10.6",
-    "axios": "^1.6.5",
+    "axios": "^1.6.7",
     "bootstrap": "5.2.2",
     "classnames": "2.5.1",
     "fast-json-patch": "^3.1.1",
